Change row count type of RowCountPrecondition from int to long (#1549)
To avoid integer overflow when tables contain more than Integer.MAX_VALUE rows, the types of `expectedRows` and `result` where changed from int/Integer to long/Long.

+++ liquibase-standard/src/main/java/liquibase/precondition/core/TableIsEmptyPrecondition.java
@@ -3,11 +3,11 @@ package liquibase.precondition.core;
 public class TableIsEmptyPrecondition extends RowCountPrecondition {
 
     public TableIsEmptyPrecondition() {
-        this.setExpectedRows(0);
+        this.setExpectedRows(0L);
     }
 
     @Override
-    protected String getFailureMessage(int result, int expectedRows) {
+    protected String getFailureMessage(long result, long expectedRows) {
         return "Table "+getTableName()+" is not empty. Contains "+result+" rows";
     }
